+++ packages/server/graphql/resolvers/addMenuItem.mutation.js
@@ -1,16 +1,11 @@
-import { PrismaClient } from "@prisma/client";
-const prisma = new PrismaClient();
-
-const addMenuItem = async (parent, args) => {
+const addMenuItem = async (parent, args, { prisma }) => {
   const newMenuItem = args.menuItem;
-  prisma.menuItem
-    .create({
-      data: newMenuItem,
-    })
-    .then((response) => {
-      console.log({ response });
-      return response;
-    });
+  const response = await prisma.menuItem.create({
+    data: newMenuItem,
+  });
+
+  console.log({ response });
+  return response;
 };
 
 export default addMenuItem;
