+++ packages/server/graphql/schema.js
@@ -1,5 +1,5 @@
 import { createSchema } from "graphql-yoga";
-import typeDefs from "./typeDefs.js";
+import typeDefs from "./typeDefs/index.js";
 import resolvers from "./resolvers/index.js";
 
 const schema = createSchema({
