+++ packages/server/package.json
@@ -11,9 +11,11 @@
     "express": "^4.18.3",
     "express-graphql": "^0.12.0",
     "graphql": "^16.8.1",
+    "graphql-subscriptions": "^2.0.0",
     "graphql-yoga": "^5.1.1",
     "moment": "^2.30.1",
-    "mongodb": "^6.4.0"
+    "mongodb": "^6.4.0",
+    "typed-graphql-subscriptions": "^0.0.1"
   },
   "devDependencies": {
     "nodemon": "^3.1.0",
