+++ packages/server/prisma/schema.prisma
@@ -15,7 +15,7 @@ datasource db {
 
 model MenuItem {
   id String @id @default(auto()) @map("_id") @db.ObjectId
-  name String @unique
+  name String
   type String
   description String
   basePrice Float
